Updating SimpleITK version to improve compatibility with Brainsight
This solves a rare occasional issue that results in files not being able to enforce isotropic conditions with SimpleITK (and using flirt as backup). Uisng SimpleITK seems to work for the majority of cases, but for the rare cases when flirt is used, it created a file that seemed to be ok (loadable with fsleyes, 3D slicer, etc) but NO with Brainsight.  Upgrading to SimpleITK 2.3.1 seems to correct the problem.